requirements.txt
youtube-transcript-api
requests
google-generativeai
python-dotenv